Menu restaurante

+++ app/src/main/java/edu/uoc/easyorderfront/domain/model/Order.kt
@@ -5,7 +5,7 @@ import edu.uoc.easyorderfront.data.order.OrderDTO
 
 class Order(
         val uid: String?,
-        val price: Int? = 0,
+        val price: Double? = 0.0,
         val state: String? = null,
         val startedTime: Long? = null,
         val orderedDishes: MutableList<OrderedDish>? = ArrayList()
