+++ app/src/main/java/edu/uoc/easyorderfront/domain/model/OrderedDish.kt
@@ -5,7 +5,7 @@ import edu.uoc.easyorderfront.data.order.OrderedDishDTO
 class OrderedDish(
         val uid: String?,
         val quantity: Int? = 0,
-        val totalPrice: Int? = 0
+        val totalPrice: Double? = 0.0
         //val dish: Dish?
 ) {
     fun converToDTO() : OrderedDishDTO {
